Move connection-management.hh to shared place.
